backend/attendance/permissions.py
# from rest_framework.permissions import BasePermission

# class IsAdmin(BasePermission):
#     def has_permission(self, request, view):
#         return hasattr(request.user, 'role') and request.user.role == 'superadmin'

# attendance/permissions.py
from rest_framework.permissions import BasePermission

class IsAdmin(BasePermission):
    def has_permission(self, request, view):
        return bool(request.user and request.user.is_authenticated and request.user.admin_role in ['admin', 'superadmin'])